Add option to configure css of individual columns
Add option to configure css of individual columns

Configured via a column_style config setting, which contains a array of the columns to style, which themselves contain an array of css elements and their values:

example:
column_style:
  2: 
    max-width: none

Have aimed for a simpler, but possibly more confusing, configuration system, and simpler code. Might be better to have more wordy configuration (column: 2, style: "max-width: none") to reduce user confusion, but that requires more processing in the code, and so more potential complex problems.